+++ foxes/utils/wrg_utils.py
@@ -1,14 +1,33 @@
 import pandas as pd
+from pathlib import Path
 
-class ReaderWRG():
 
-    def __init__(self, file_name):
-        self._file_name = file_name
+class ReaderWRG:
+    """
+    A reader for WRG files
+
+    Attributes
+    ----------
+    fpath: pathlib.Path
+        Path to the wrg file
+
+    """
+
+    def __init__(self, fpath):
+        """
+        Constructor
+        
+        Parameters
+        ----------
+        fpath: str
+            Path to the wrg file
+        """
+        self.fpath = Path(fpath)
         self._prepare()
 
     def _prepare(self):
         # read the first two lines
-        with open(self._file_name) as fstream:
+        with open(self.fpath, "r") as fstream:
             self._nx, self._ny, self._utmx0, self._utmy0, self._res = fstream.readline().split()
             second_line = fstream.readline().split()
             n_cols = len(second_line)
@@ -31,12 +50,22 @@ class ReaderWRG():
         cols[9::3] = cols_sel("As", self._n_sectors)
         cols[10::3] = cols_sel("Ks", self._n_sectors)
 
-        self._data = pd.read_csv(self._file_name, names=cols, skiprows=1,
+        self._data = pd.read_csv(self.fpath, names=cols, skiprows=1,
                                  sep='\s+', usecols=range(1, n_cols))
 
         self._data[cols_sel("fs", self._n_sectors)] /= 10
         self._data[cols_sel("As", self._n_sectors)] /= 10
         self._data[cols_sel("Ks", self._n_sectors)] /= 100
 
+    @property
     def data(self):
+        """
+        The WRG data
+        
+        Returns
+        -------
+        df: pandas.DataFrame
+            The WRG data
+        
+        """
         return self._data